requirements.txt
pdf2image==1.17.0
Pillow==10.2.0
setuptools==69.5.1
customtkinter==5.2.2
